+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix AltGr / Right option detection on Mac OS (#206)
 - Use window_decorations setting for settings UI and about (#297)
 - Fix incorrect column display when indenting with tabs (#289)
+- Support /path/to/file.txt(ln, col) style paths in terminal (second attempt) (#75)
 
 [sumi-light]: https://github.com/LogicSatinn/sumi-light-zed